+++ src/components/ColourSwatch.vue
@@ -42,7 +42,6 @@ import { reactive, computed } from "vue";
 import FormAction from "@/components/FormAction.vue";
 import IconDustbin from "@/components/icons/IconDustbin.vue";
 import IconCopy from "@/components/icons/IconCopy.vue";
-import hexToRGB from "@/composables/hexToRGB.js";
 import { useColourStore } from "@/stores/colourStore";
 const colourStore = useColourStore();