+++ src/composables/hexToRGB.js
@@ -9,7 +9,7 @@ export default function hexToRGB(hex) {
   rgbArray.push(h >>> (alpha ? 24 : 16));
   rgbArray.push((h & (alpha ? 0x00ff0000 : 0x00ff00)) >>> (alpha ? 16 : 8));
   rgbArray.push((h & (alpha ? 0x0000ff00 : 0x0000ff)) >>> (alpha ? 8 : 0));
-  rgbArray.push(alpha ? `, ${h & 0x000000ff}` : 255);
+  rgbArray.push(alpha ? (h & 0x000000ff) : 255);
 
   return rgbArray;
 }